templates/requirements.txt
pyyaml>=6.0
pytest>=7.0
pytest-mock>=3.10
pytest-timeout>=2.1
